nosql/vector/requirements.txt
faiss-cpu
numpy